New folder/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-google-genai
langchain-community 
pytesseract
opencv-python
pillow